Added DeathContextualizer
It uses HarmonyX to perform a patch that searches for the Creature.Die method and calls an event beforehand.
Shoutout to EtiTheSpirit for originally demonstrating this in their mod.

Death messages for Centipedes, Rain, ZapCoils, Worm Grass, Five Pebbles, Wall Rot and Player Tounge are now processed

Invalid deaths are no longer processed