feat: dynamically positioned slugcat portraits
Slugcat select portraits are split evenly between two layers, if there is an odd number of slugcats the bottom row takes the extra portrait.

Have not tested with an odd number of top portraits (only achievable in modded scenarios) but it uses the same logic as the bottom row to account for placement so it probably works.